+++ scico/optimize/_padmm.py
